tokio: move I/O helpers to ext traits
Refs: #1203
